Refactoring: routes and apicontroller
- Removed ForemanLeapp namespaces
- Renamed POST method to use default routing
- Used namespaces instead of hardcoded api/v2
- Preupgrade template refactored to accommodate
  changes mentioned above.
