+++ app/models/preupgrade_report_entry.rb
@@ -1,0 +1,6 @@
+class PreupgradeReportEntry < ApplicationRecord
+  belongs_to :message
+  belongs_to :preupgrade_report
+
+  validates :message_id, :preupgrade_report_id, :presence => true
+end
